Add files via upload
Web scraping is a technique to extract and parse the data from websites in an automated fashion using the computer programming language. In this project,
requests package was used for pulling the data from the webpage source and import requests module used to retrieve the data from HTML or XML files.
Beautiful Soup is used to parse and extract data from Latin NCAP website which tests different cars on their safety measures.
In order to depict the data in the form of tabular sheets and that's why pandas library has been installed.
However ,the data was not available in an excel sheet that's why web scraping has to be done to do further analysis regarding the safest cars.
Consequently, Datas for 2020-2021 have been extracted because of different formats followed by the remaining years.